requirements.txt
alembic==1.7.6
altair==4.2.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
bson==0.5.10
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.0.0
conda==4.3.16
cycler==0.11.0
databricks-cli==0.16.4
dataflow==0.9.5
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.0
docker==5.0.3
dominate==2.6.0
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
fonttools==4.29.1
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==20.1.0
huggingface-hub==0.4.0
idna==3.3
importlib-metadata==4.11.2
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonlines==3.0.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.8.0
Mako==1.2.0
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mlflow==1.24.0
mlflow-skinny==1.24.0
msgpack==1.0.3
msgpack-numpy==0.4.7.1
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.7.1
nltk==3.7
notebook==6.4.8
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pycocotools==2.0.4
pycosat==0.6.3
pycparser==2.21
pydeck==0.7.1
Pygments==2.11.2
pymongo==4.0.2
Pympler==1.0.1
pyparsing==3.0.7
PyPDF2==1.26.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
regex==2022.3.2
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.7.0
tabulate==0.8.9
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
threadpoolctl==3.1.0
tokenizers==0.11.6
toml==0.10.2
toolz==0.11.2
torch==1.10.2
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
tweepy==4.6.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
validators==0.18.2
visitor==0.1.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.5.2
wordcloud==1.8.1
zipp==3.7.0
